backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi-jwt-auth==0.5.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
platformdirs==4.3.6
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==1.7.1
pymongo==4.10.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.0
virtualenv==20.28.1
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.1.3
